unify the user mode audit entry in rust too
