+++ proxy_agent/src/redirector/linux.rs
@@ -1,7 +1,5 @@
 // Copyright (c) Microsoft Corporation
 // SPDX-License-Identifier: MIT
-mod ebpf_obj;
-
 use crate::common::{
     config, constants,
     error::{BpfErrorType, Error},
@@ -16,7 +14,7 @@ use aya::{
     programs::CgroupAttachMode,
 };
 use aya::{Btf, Ebpf, EbpfLoader};
-use ebpf_obj::{
+use crate::redirector::shared_ebpf::linux_types::{
     destination_entry, sock_addr_audit_entry, sock_addr_audit_key, sock_addr_skip_process_entry,
 };
 use proxy_agent_shared::telemetry::event_logger;
@@ -488,8 +486,9 @@ pub async fn update_hostga_redirect_policy(
 mod tests {
     use crate::common::config;
     use crate::common::constants;
-    use crate::redirector::linux::ebpf_obj::sock_addr_audit_entry;
-    use crate::redirector::linux::ebpf_obj::sock_addr_audit_key;
+    use crate::redirector::shared_ebpf::linux_types::{
+        sock_addr_audit_entry, sock_addr_audit_key,
+    };
     use aya::maps::HashMap;
     use proxy_agent_shared::misc_helpers;
     use std::env;
