+++ proxy_agent/src/redirector/windows/bpf_obj.rs
@@ -2,12 +2,12 @@
 // SPDX-License-Identifier: MIT
 #![allow(non_camel_case_types)]
 
-use crate::common::{
-    error::{BpfErrorType, Error},
-    result::Result,
-};
 use std::ffi::c_char;
-use std::ffi::c_void;
+
+pub use crate::redirector::shared_ebpf::windows_types::{
+    destination_entry_t, sock_addr_audit_key_t, sock_addr_skip_process_entry,
+};
+pub(crate) use crate::redirector::shared_ebpf::AuditValueEntry;
 
 pub type ebpf_id_t = u32;
 pub type fd_t = i32;
@@ -226,269 +226,6 @@ pub struct bpf_object {
     execution_type: ebpf_execution_type_t,
 }
 
-#[repr(C)]
-pub struct _sock_addr_audit_key {
-    pub protocol: u32,
-    pub source_port: u32,
-}
-pub type sock_addr_audit_key_t = _sock_addr_audit_key;
-impl sock_addr_audit_key_t {
-    pub fn from_source_port(port: u16) -> Self {
-        sock_addr_audit_key_t {
-            protocol: IPPROTO_TCP,
-            source_port: u32::from(port.to_be()),
-        }
-    }
-}
-
-#[repr(C)]
-pub struct _sock_addr_audit_entry {
-    pub logon_id: u32,
-    pub process_id: u32,
-    pub is_root: u32,
-    pub destination_ipv4: u32,
-    pub destination_port: u32,
-}
-pub type sock_addr_audit_entry_t = _sock_addr_audit_entry;
-impl sock_addr_audit_entry_t {
-    pub fn empty() -> Self {
-        sock_addr_audit_entry_t {
-            logon_id: 0,
-            process_id: 0,
-            is_root: 0,
-            destination_ipv4: 0,
-            destination_port: 0,
-        }
-    }
-
-    pub fn to_audit_entry(&self) -> crate::redirector::AuditEntry {
-        crate::redirector::AuditEntry {
-            logon_id: self.logon_id as u64,
-            process_id: self.process_id,
-            is_admin: self.is_root as i32,
-            destination_ipv4: self.destination_ipv4,
-            destination_port: self.destination_port as u16,
-        }
-    }
-}
-
 pub fn bpf_map_value_size(map: *const bpf_map) -> u32 {
     unsafe { (*map).map_definition.value_size }
 }
-
-// Legacy layout used by older Windows eBPF programs before shared struct migration.
-#[repr(C)]
-pub struct _sock_addr_audit_entry_legacy {
-    pub logon_id: u64,
-    pub process_id: u32,
-    pub is_admin: i32,
-    pub destination_ipv4: u32,
-    pub destination_port: u16,
-}
-pub type sock_addr_audit_entry_legacy_t = _sock_addr_audit_entry_legacy;
-impl sock_addr_audit_entry_legacy_t {
-    pub fn empty() -> Self {
-        sock_addr_audit_entry_legacy_t {
-            logon_id: 0,
-            process_id: 0,
-            is_admin: 0,
-            destination_ipv4: 0,
-            destination_port: 0,
-        }
-    }
-
-    pub fn to_audit_entry(&self) -> crate::redirector::AuditEntry {
-        crate::redirector::AuditEntry {
-            logon_id: self.logon_id,
-            process_id: self.process_id,
-            is_admin: self.is_admin,
-            destination_ipv4: self.destination_ipv4,
-            destination_port: self.destination_port,
-        }
-    }
-}
-
-/// Represents an unknown audit value entry.
-pub struct GenericAuditValueEntry {
-    /// Owned bytes for unknown layout values.
-    ///
-    /// We keep this as raw bytes and decode based on `read_size` to avoid
-    /// assuming a concrete Rust type up front.
-    pub buffer: Vec<u8>,
-    /// Allocated size of the audit value data.
-    pub data_size: u32,
-    /// Number of bytes read from the unknown audit value data.
-    pub read_size: u32,
-}
-
-pub(crate) enum AuditValueEntry {
-    New(sock_addr_audit_entry_t),
-    Legacy(sock_addr_audit_entry_legacy_t),
-    // if Unknown type, use the generic audit value entry
-    Unknown(GenericAuditValueEntry),
-}
-
-impl AuditValueEntry {
-    const VALUE_SIZE_NEW: u32 = std::mem::size_of::<sock_addr_audit_entry_t>() as u32;
-    const VALUE_SIZE_LEGACY: u32 = std::mem::size_of::<sock_addr_audit_entry_legacy_t>() as u32;
-
-    /// return the maximum value size among all audit entry types.
-    fn max_value_size() -> u32 {
-        Self::VALUE_SIZE_NEW.max(Self::VALUE_SIZE_LEGACY)
-    }
-
-    pub fn empty(value_size: u32) -> Self {
-        match value_size {
-            Self::VALUE_SIZE_NEW => AuditValueEntry::New(sock_addr_audit_entry_t::empty()),
-            Self::VALUE_SIZE_LEGACY => {
-                AuditValueEntry::Legacy(sock_addr_audit_entry_legacy_t::empty())
-            }
-            _ => {
-                let max_value_size = Self::max_value_size();
-                AuditValueEntry::Unknown(GenericAuditValueEntry {
-                    // Allocate empty buffer with the maximum value size we currently supported.
-                    buffer: vec![0u8; max_value_size as usize],
-                    data_size: max_value_size,
-                    // For map lookups, value_size is fixed per map and equals the
-                    // number of bytes written by bpf_map_lookup_elem.
-                    read_size: value_size,
-                })
-            }
-        }
-    }
-
-    pub fn value_size(&self) -> u32 {
-        match self {
-            AuditValueEntry::New(_) => Self::VALUE_SIZE_NEW,
-            AuditValueEntry::Legacy(_) => Self::VALUE_SIZE_LEGACY,
-            AuditValueEntry::Unknown(entry) => entry.data_size,
-        }
-    }
-
-    pub fn value_pointer_mut(&mut self) -> *mut c_void {
-        match self {
-            AuditValueEntry::New(entry) => entry as *mut sock_addr_audit_entry_t as *mut c_void,
-            AuditValueEntry::Legacy(entry) => {
-                entry as *mut sock_addr_audit_entry_legacy_t as *mut c_void
-            }
-            AuditValueEntry::Unknown(entry) => entry.buffer.as_mut_ptr() as *mut c_void,
-        }
-    }
-
-    pub fn to_audit_entry(&self) -> Result<crate::redirector::AuditEntry> {
-        match self {
-            AuditValueEntry::New(entry) => Ok(entry.to_audit_entry()),
-            AuditValueEntry::Legacy(entry) => Ok(entry.to_audit_entry()),
-            AuditValueEntry::Unknown(entry) => {
-                // Cast the raw bytes to the concrete audit value type based on
-                // the number of bytes read.
-                match entry.read_size {
-                    Self::VALUE_SIZE_NEW => {
-                        if entry.buffer.len() < Self::VALUE_SIZE_NEW as usize {
-                            return Err(Error::Bpf(BpfErrorType::MapLookupElem(
-                                "into_audit_entry".to_string(),
-                                format!(
-                                    "Insufficient buffer length {} for canonical audit entry size {}",
-                                    entry.buffer.len(),
-                                    Self::VALUE_SIZE_NEW
-                                ),
-                            )));
-                        }
-
-                        let value = unsafe {
-                            std::ptr::read_unaligned(
-                                entry.buffer.as_ptr() as *const sock_addr_audit_entry_t
-                            )
-                        };
-                        Ok(value.to_audit_entry())
-                    }
-                    Self::VALUE_SIZE_LEGACY => {
-                        if entry.buffer.len() < Self::VALUE_SIZE_LEGACY as usize {
-                            return Err(Error::Bpf(BpfErrorType::MapLookupElem(
-                                "into_audit_entry".to_string(),
-                                format!(
-                                    "Insufficient buffer length {} for legacy audit entry size {}",
-                                    entry.buffer.len(),
-                                    Self::VALUE_SIZE_LEGACY
-                                ),
-                            )));
-                        }
-
-                        let value = unsafe {
-                            std::ptr::read_unaligned(
-                                entry.buffer.as_ptr() as *const sock_addr_audit_entry_legacy_t
-                            )
-                        };
-                        Ok(value.to_audit_entry())
-                    }
-                    _ => Err(Error::Bpf(BpfErrorType::MapLookupElem(
-                        "into_audit_entry".to_string(),
-                        format!(
-                            "Invalid audit value size read: {}, allocated: {}",
-                            entry.read_size, entry.data_size
-                        ),
-                    ))),
-                }
-            }
-        }
-    }
-}
-
-#[repr(C)]
-pub struct _ip_address {
-    //pub ipv4: u32,
-    pub ip: [u32; 4], // ipv4 uses the first element; ipv6 uses all 4 elements
-}
-impl _ip_address {
-    fn empty() -> Self {
-        _ip_address { ip: [0, 0, 0, 0] }
-    }
-
-    pub fn from_ipv4(ipv4: u32) -> Self {
-        let mut ip = Self::empty();
-        ip.ip[0] = ipv4;
-        ip
-    }
-
-    #[allow(dead_code)]
-    pub fn from_ipv6(ipv6: [u32; 4]) -> Self {
-        let mut ip = Self::empty();
-        ip.ip.copy_from_slice(&ipv6);
-        ip
-    }
-}
-pub type ip_address_t = _ip_address;
-
-#[repr(C)]
-pub struct _destination_entry {
-    pub destination_ip: ip_address_t,
-    pub destination_port: u32,
-    pub protocol: u32,
-}
-impl _destination_entry {
-    pub fn empty() -> Self {
-        _destination_entry {
-            destination_ip: ip_address_t::empty(),
-            destination_port: 0,
-            protocol: IPPROTO_TCP,
-        }
-    }
-
-    pub fn from_ipv4(ipv4: u32, port: u16) -> Self {
-        let mut entry = Self::empty();
-        entry.destination_ip = ip_address_t::from_ipv4(ipv4);
-        entry.destination_port = u32::from(port.to_be());
-        entry
-    }
-}
-
-pub type destination_entry_t = _destination_entry;
-pub const IPPROTO_TCP: u32 = 6;
-#[allow(dead_code)]
-pub const IPPROTO_UDP: u32 = 17;
-
-#[repr(C)]
-pub struct _sock_addr_skip_process_entry {
-    pub pid: u32,
-}
-pub type sock_addr_skip_process_entry = _sock_addr_skip_process_entry;
